// src/core/payments/dto/create-payment.dto.ts
export class PaymentByMomoParam {
  partnerCode: string;
  partnerName: string;
  storeId: string;
  requestType: string;
  ipnUrl: string;
  redirectUrl: string;
  orderId: string;
  amount: number;
  lang: string;
  orderInfo: string;
  requestId: string;
  extraData: string;
  signature: string;
}